requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
neo4j>=5.14.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.3.0
